back out changes to JestClientFactory and JestHttpClientTest now that signer is structured as interceptor
